requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.18
django-jalali-date==1.1.3
django-render-partial==0.4
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
jalali_core==1.0.0
jdatetime==5.2.0
lxml==5.3.0
multidict==6.1.0
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
pytz==2025.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
zeep==4.3.1
